fix: Tira a responsabilidade do model ninja e passa para o DTO nas requisições

+++ src/main/java/com/kava/s/CadastroNinja/ninja/service/NinjaService.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class NinjaService {
@@ -28,29 +29,40 @@ public class NinjaService {
     }
 
     //GET -- Manda uma requisição para listar todos os ninjas
-    public List<NinjaModel> listarNinjas(){
-        return ninjaRepository.findAll();
+    public List<NinjaDTO> listarNinjas(){
+        List<NinjaModel> ninjaModel = ninjaRepository.findAll();
+
+        return ninjaModel.stream()
+                .map(ninjaMapper::map)
+                .collect(Collectors.toList());
     }
 
     //GET -- Manda uma requisição para listar ninja por ID
-    public NinjaModel listarNinjaId(Long id){
+    public NinjaDTO listarNinjaId(Long id){
         Optional<NinjaModel> caixaNinja = ninjaRepository.findById(id);
 
-        return caixaNinja.orElse(null);
+        return caixaNinja.map(ninjaMapper::map).orElse(null);
     }
 
     //PUT -- Manda uma requisição para alterar um ninja
-    public NinjaModel alterarNinja(Long id, NinjaModel ninjaAtualizado){
-        if(ninjaRepository.existsById(id)){
-            ninjaAtualizado.setId(id);
-            return ninjaRepository.save(ninjaAtualizado);
-        }
+    public NinjaDTO alterarNinja(Long id, NinjaDTO ninjaDTO){
+        Optional<NinjaModel> caixaNinja = ninjaRepository.findById(id);
 
+        if(caixaNinja.isPresent()){
+            NinjaModel ninja = ninjaMapper.map(ninjaDTO);
+            ninja.setId(id);
+
+            return ninjaMapper.map(ninja);
+        }
         return null;
     }
 
     //DELETE -- Manda uma requisição para deletar um ninja
     public void deletarNinja(Long id){
-        ninjaRepository.deleteById(id);
+        Optional<NinjaModel> caixaNinja = ninjaRepository.findById(id);
+
+        if(caixaNinja.isPresent()){
+            ninjaRepository.deleteById(id);
+        }
     }
 }
